requirements.txt
flask
flask-cors
requests
beautifulsoup4
sendgrid
openai
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
python-dotenv
APScheduler
supabase
gunicorn
